homework2/backend/pyproject.toml
[project]
name = "coding-interview-backend"
version = "0.1.0"
description = "Backend for collaborative coding interview platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-socketio>=5.10.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "alembic>=1.12.1",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# For a flat project structure without a package directory
# We need to explicitly list all files to include
# This tells hatchling exactly which Python files to package
only-include = [
    "__init__.py",
    "main.py",
    "database.py",
    "db_service.py",
    "code_executor.py",
    "migrations.py",
]

